Provide detailed feedback.

Function formatting the feedback about any type of questions into a HTML table which can be embedded in an e-mail to be sent to the selected student.

provide_detailed_feeback(
  details,
  studentid,
  labels = base::list(values = c("Wrong", "Missing", "Partial", "Correct"), table_names =
    c("Proposition:", "This proposition is:", "Therefore your answer is:",
    "And you earned:", "Explanation", "Topic"))
)

Arguments

details

Tibble. Feedback recorded for all items of all versions of all questions.

studentid

Character. ID of the student.

labels

List. Names for axes and legends (for translations purposes)

Value

HTML code of the Individual feedback properly formatted in a table.

Author

Nicolas Mangin